chore: Implement `useGetAreas` hook
